move upload logic from add contacts to all contacts grid
